Backend

See merge request s08-webmobile1-sub2/S08P12E104!178

+++ backend/src/main/java/com/ssafy/patpat/shelter/dto/RequestShelterDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RequestShelterDto {
-    private int breedId;
+    private Long breedId;
     private String sidoCode;
     private String gugunCode;
-    private int offSet;
-    private int limit;
+    private Integer offSet;
+    private Integer limit;
 }
