+++ backend/src/main/java/com/ssafy/patpat/user/dto/UserDto.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.Size;
 @Builder
 public class UserDto {
 
-    private long userId;
+    private Long userId;
     @Size(min=1, max=50)
     private String provider;
 
@@ -27,7 +27,7 @@ public class UserDto {
 
     private String providerId;
 
-    private Integer shelterId;
+    private Long shelterId;
 
 
 }
